moving class to database_utils file instead of data_extraction file
